requirements.txt
--extra-index-url https://pypi.org/simple/
google-cloud-dlp
google-cloud-kms
apache-beam[gcp]
